requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.10.0
semantic-kernel==0.9.0
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
flake8==6.1.0
black==23.12.1
